requirements.txt
# Core Data Processing
pandas>=2.1.0
numpy>=1.24.0
pandas-ta>=0.3.14b

# Database
psycopg2-binary>=2.9.7

# Environment & Configuration
python-dotenv>=1.0.0
pytz>=2023.3

# API Integration
requests>=2.31.0
kiteconnect>=4.2.0

# System Monitoring
psutil>=5.9.0

# Scientific Computing
scipy>=1.11.0
scikit-learn>=1.3.0

# Utilities
openpyxl>=3.1.0
schedule>=1.2.0

# Optional: Telegram notifications
python-telegram-bot>=20.0